Move BenchmarkRunner to the tests package
