+++ integration_tests/src/main/scala/com/nvidia/spark/rapids/tests/tpcds/TpcdsLikeBench.scala
@@ -16,7 +16,7 @@
 
 package com.nvidia.spark.rapids.tests.tpcds
 
-import com.nvidia.spark.rapids.tests.common.{BenchmarkRunner, BenchmarkSuite}
+import com.nvidia.spark.rapids.tests.common.BenchmarkSuite
 
 import org.apache.spark.sql.{DataFrame, SparkSession}
